% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructed_model}
\alias{constructed_model}
\title{Constructsa new model with noised residuals
y_new  = y_fitted + residuals*noise}
\usage{
constructed_model(
  fitted_dep_var,
  residuals,
  data_indep,
  distribution = "rnorm"
)
}
\arguments{
\item{fitted_dep_var}{fitted values of an lm model}

\item{residuals}{residuals of a model}

\item{data_indep}{independent data used to adjust an lm model}

\item{distribution}{Type of noise added to residuals, ej "rnorm" or "rrademacher"}
}
\value{
Constructed lm model
}
\description{
Constructsa new model with noised residuals
y_new  = y_fitted + residuals*noise
}
\examples{
y_hat <- 1:10
residuals <- rnorm(10)
data_indep <- data.frame(x=10:19)
constructed_model(y_hat, residuals, data_indep)
}
