% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals_statistics.R
\name{cvm_value}
\alias{cvm_value}
\title{Calculates the Cramer von Mises value
given an lm model with its residuals.}
\usage{
cvm_value(model)
}
\arguments{
\item{model}{An lm model.}
}
\value{
The Cramer von Mises value of the model.
}
\description{
Calculates the Cramer von Mises value
given an lm model with its residuals.
}
\examples{
x <- 1:10
y <- 2*x + rnorm(10)
model <- lm(y~x-1)
cvm_value(model)
}
