% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{dominguez_lobato_test}
\alias{dominguez_lobato_test}
\title{Tests the specification of a linear model using wild bootrap}
\usage{
dominguez_lobato_test(
  model,
  times = 300,
  distribution = "rnorm",
  statistic = "cvm_value",
  verbose = FALSE,
  quantiles = c(0.9, 0.95, 0.99)
)
}
\arguments{
\item{model}{An lm model}

\item{times}{Number of bootstrap samples}

\item{distribution}{Type of noise added to residuals, ej "rnorm" or "rrademacher"}

\item{statistic}{Type of statistic to be used, can be one of "cmv_value" or "kmv_value"}

\item{verbose}{TRUE for print each bootstrap iteration}

\item{quantiles}{vector of quantiles to calculate pvalues}
}
\value{
A list with a data.frame results and the orderend values of each boostrap iteration
}
\description{
Tests the specification of a linear model using wild bootrap
}
\examples{
x <- 1:10
y <- 1:10
model <- lm(y~x-1)
dominguez_lobato_test(model)
}
\references{
Manuel A. Domínguez and Ignacio N. Lobato (2019)
*Specification testing with estimated variables.* Econometrics Reviews.
}
