% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{independent_data}
\alias{independent_data}
\title{Gets the independent data of a lm model}
\usage{
independent_data(model)
}
\arguments{
\item{model}{An lm model}
}
\value{
A dataframe with the independet data of the lm model
}
\description{
Gets the independent data of a lm model
}
\examples{
x <- 1:10
z <- x**2
y <- 1:10
model <- lm(y~x+z)
independent_data(model)
}
