% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals_statistics.R
\name{kmv_value}
\alias{kmv_value}
\title{Calculates the Kolmogorov value
given an lm model with its residuals}
\usage{
kmv_value(model)
}
\arguments{
\item{model}{A lm model}
}
\value{
The Kolmogorov value of the model
}
\description{
Calculates the Kolmogorov value
given an lm model with its residuals
}
\examples{
x <- 1:10
y <- 2*x + rnorm(10)
model <- lm(y~x-1)
kmv_value(model)
}
