% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_dl_test}
\alias{plot_dl_test}
\title{Plots the Dominguez-Lobato test.}
\usage{
plot_dl_test(x)
}
\arguments{
\item{x}{An object of class `dl_test`.}
}
\value{
Plot of type ggplot.
}
\description{
Plots the Dominguez-Lobato test.
}
\examples{
x <- 1:10
y <- 1:10
model <- lm(y~x-1)
dl_test <- dominguez_lobato_test(model)
plot_dl_test(dl_test)
}
