% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_reset_test}
\alias{plot_reset_test}
\title{Plot the reset test.}
\usage{
plot_reset_test(x)
}
\arguments{
\item{x}{An object of class `reset_test`.}
}
\value{
Plot of type ggplot.
}
\description{
Plot the reset test.
}
\examples{
x <- 1:10
y <- 1:10
model <- lm(y~x-1)
r_test <- reset_test(model)
plot_reset_test(r_test)
}
