% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions.R
\name{rmammen_cont}
\alias{rmammen_cont}
\title{Random deviates of Mammen continuous distribution.}
\usage{
rmammen_cont(n)
}
\arguments{
\item{n}{Number of observations.}
}
\value{
Random deviates of size n.
}
\description{
Random deviates of Mammen continuous distribution.
}
\examples{
rmammen_cont(10)
}
