% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions.R
\name{rmammen_point}
\alias{rmammen_point}
\title{Random deviates of Mammen distribution.}
\usage{
rmammen_point(n)
}
\arguments{
\item{n}{Number of observations.}
}
\value{
Random deviates of size n.
}
\description{
Random deviates of Mammen distribution.
}
\examples{
rmammen_point(10)
}
