% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions.R
\name{rrademacher}
\alias{rrademacher}
\title{Random deviates of Rademacher distribution.}
\usage{
rrademacher(n)
}
\arguments{
\item{n}{Number of observations.}
}
\value{
Random deviates of size n.
}
\description{
Random deviates of Rademacher distribution.
}
\examples{
rrademacher(10)
}
