% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals_statistics.R
\name{presiduals}
\alias{presiduals}
\title{Calculates the accumulated distribution of
residuals at each residual point}
\usage{
presiduals(fitted_values, resids)
}
\arguments{
\item{fitted_values}{vector of fitted values}

\item{resids}{residuals vector of each fitted value}
}
\value{
vector of size len(resids)
}
\description{
Calculates the accumulated distribution of
residuals at each residual point
}
\examples{
y_hat <- c(4, 8, 7)
resids <- c(1, 5, 3)
presiduals(y_hat, resids)
}
