% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions.R
\name{rmammen_cont}
\alias{rmammen_cont}
\title{Random deviates of mammen continouos distribution.}
\usage{
rmammen_cont(n)
}
\arguments{
\item{n}{Number of obervations}
}
\value{
random deviates of size n
}
\description{
Random deviates of mammen continouos distribution.
}
\examples{
rmammen_cont(10)
}
