% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions.R
\name{rmammen_point}
\alias{rmammen_point}
\title{Random deviates of mammen distribution.}
\usage{
rmammen_point(n)
}
\arguments{
\item{n}{Number of obervations}
}
\value{
random deviates of size n
}
\description{
Random deviates of mammen distribution.
}
\examples{
rmammen_point(10)
}
