% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_distributions.R
\name{rrademacher}
\alias{rrademacher}
\title{Random deviates of rademacher distribution.}
\usage{
rrademacher(n)
}
\arguments{
\item{n}{Number of obervations.}
}
\value{
Random deviates of size n.
}
\description{
Random deviates of rademacher distribution.
}
\examples{
rrademacher(10)
}
