% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{wald_test}
\alias{wald_test}
\title{Custom Wald test.
Tests restrictions*coefficients = value.}
\usage{
wald_test(model, restrictions, value, robust = F)
}
\arguments{
\item{model}{An lm model.}

\item{restrictions}{Matrix of size (number of restrictions) times length(coefficients),
for free restrictions use zeros.}

\item{value}{Values of restrictions.}

\item{robust}{Use robust varcov matrix.}
}
\value{
A list with the Wald value and the corresponding pvalue.
}
\description{
Custom Wald test.
Tests restrictions*coefficients = value.
}
\examples{
x <- 1:10
z <- x**2
y <- 1:10
model <- lm(y~x+z)
restrictions <- diag(3)
value <-  as.matrix(c(0, 0, 0))
wald_test(model, restrictions, value)
}
